Update Windows uv installation docs
The docs contain two commands to run on Windows.  The first command can 
be run on any shell; the second is only valid inside PowerShell.  While 
the syntax highlighting hints already indicate this, there is no way for
the end user to have any knowledge of this; so this commit adds a 
notation on the second command denoting it should be run inside 
PowerShell