Fix TTS settings formatting in documentation
The documentation has two settings fields reversed when compared to Open WebUI actual settings screen. This causes some people (like me) to input the model in the voice box and the other way around, causing strange errors. By swapping the two fields in the docs, we follow natural order of information (top to bottom, left to right) and make the documentation match the actual settings screen.